requirements.txt
psycopg[binary]~=3.1.9
python-dotenv~=1.0.0
Flask~=2.3.2
Flask-JWT-Extended~=4.4.4
requests~=2.31.0
argon2-cffi~=21.3.0
email-validator~=2.0.0
